refactor(SecurityConfig): 패키지 구조 리펙토링
