+++ src/main/java/com/nhnacademy/front/account/member/adaptor/MemberLoginAdaptor.java
@@ -1,11 +1,11 @@
-package com.nhnacademy.front.account.member.login.adaptor;
+package com.nhnacademy.front.account.member.adaptor;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import com.nhnacademy.front.account.member.login.model.dto.request.RequestLoginMemberDTO;
-import com.nhnacademy.front.account.member.login.model.dto.response.ResponseLoginMemberDTO;
+import com.nhnacademy.front.account.member.model.dto.request.RequestLoginMemberDTO;
+import com.nhnacademy.front.account.member.model.dto.response.ResponseLoginMemberDTO;
 
 @FeignClient(name = "member-login-user-details-service", url = "http://e2bs.shop/api/login")
 public interface MemberLoginAdaptor {
