+++ src/main/java/com/nhnacademy/front/account/member/controller/MemberLoginController.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.account.member.login.controller;
+package com.nhnacademy.front.account.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
