+++ src/main/java/com/nhnacademy/front/account/member/exception/LoginProcessException.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.front.account.member.login.exception;
+package com.nhnacademy.front.account.member.exception;
 
 public class LoginProcessException extends RuntimeException {
 	public LoginProcessException(String message) {
