--- src/main/java/com/nhnacademy/front/account/member/login/model/domain/RankName.java
@@ -1,7 +1,0 @@
-package com.nhnacademy.front.account.member.login.model.domain;
-
-public enum RankName {
-
-	NORMAL, ROYAL, GOLD, PLATINUM
-
-}
