+++ src/main/java/com/nhnacademy/front/account/member/model/domain/RankName.java
@@ -1,0 +1,7 @@
+package com.nhnacademy.front.account.member.model.domain;
+
+public enum RankName {
+
+	NORMAL, ROYAL, GOLD, PLATINUM
+
+}
