+++ src/main/java/com/nhnacademy/front/account/member/model/dto/response/ResponseLoginMemberDTO.java
@@ -1,6 +1,6 @@
-package com.nhnacademy.front.account.member.login.model.dto.response;
+package com.nhnacademy.front.account.member.model.dto.response;
 
-import com.nhnacademy.front.account.member.login.model.domain.RankName;
+import com.nhnacademy.front.account.member.model.domain.RankName;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
